+++ Vertex2D.h
@@ -40,23 +40,23 @@ public:
     //_____________________________________________________________________________________________________________________________________
     //_____________________________________________________________________________________________________________________________________
 
-    Vector2D& add(double other); // USED FOR THE MANIPULATION OF BOTH ELEMENTS OF THE VECTOR WITH A SINGLE NUMBER
-    Vector2D& subtract(double other); // 
-    Vector2D& divide(double other); // USAGE:    VECTOR2D   = float VALUE
-    Vector2D& multiply(double other); // RESULT:   VECTOR2D.X = float VALUE
-    Vector2D& assign(double other); //           VECTOR2D.Y = float VALUE
+    Vector2D& add(float other); // USED FOR THE MANIPULATION OF BOTH ELEMENTS OF THE VECTOR WITH A SINGLE NUMBER
+    Vector2D& subtract(float other); // 
+    Vector2D& divide(float other); // USAGE:    VECTOR2D   = float VALUE
+    Vector2D& multiply(float other); // RESULT:   VECTOR2D.X = float VALUE
+    Vector2D& assign(float other); //           VECTOR2D.Y = float VALUE
 
-    Vector2D& operator =(double right);
-    Vector2D& operator +=(double other);
-    Vector2D& operator -=(double other);
-    Vector2D& operator *=(double other);
-    Vector2D& operator /=(double other);
+    Vector2D& operator =(float right);
+    Vector2D& operator +=(float other);
+    Vector2D& operator -=(float other);
+    Vector2D& operator *=(float other);
+    Vector2D& operator /=(float other);
 
 
-    friend Vector2D operator +(Vector2D left, double right);  //
-    friend Vector2D operator -(Vector2D left, double right);  //
-    friend Vector2D operator /(Vector2D left, double right);  //
-    friend Vector2D operator *(Vector2D left, double right);  //
+    friend Vector2D operator +(Vector2D left, float right);  //
+    friend Vector2D operator -(Vector2D left, float right);  //
+    friend Vector2D operator /(Vector2D left, float right);  //
+    friend Vector2D operator *(Vector2D left, float right);  //
 //_____________________________________________________________________________________________________________________________________
 //_____________________________________________________________________________________________________________________________________
 
@@ -66,6 +66,9 @@ public:
 extern inline double GetAngle(Vector2D A, Vector2D B);
 
 template <typename T>
-static inline T Squared(T x) { return x * x; }
+static inline T Squared(T x)
+{
+    return x * x;
+}
 
 extern inline Vector2D Get_Displacement(Vector2D A, Vector2D B);
